requirements.txt
uvicorn
websockets
aiohttp
httpx
qq-botpy
requests
beautifulsoup4
playwright
Pillow
numpy
pandas
matplotlib
scikit-learn
pyyaml
pytest
scalene
Pympler
objgraph
orjson
aiofiles
loguru
psutil
py-spy
websockets
loguru
redis
rich
